intial

+++ src/main/java/com/robson/pride/api/elements/BloodElement.java
@@ -23,7 +23,7 @@ import static com.robson.pride.api.skillcore.SkillCore.stackablePassiveBase;
 public interface BloodElement {
 
     ElementData DATA = new ElementData("Blood", ServerDataManager.BLOOD, ParticleRegistry.BLOOD_PARTICLE.get(), ChatFormatting.DARK_RED,
-            SoundRegistry.BLOOD_NEEDLE_IMPACT.get(), (byte) 5, SchoolRegister.BLOOD.get(), new ItemRenderingParams(new FixedRGB(255, 100, 50),
+            SoundRegistry.BLOOD_NEEDLE_IMPACT.get(), (byte) 5, SchoolRegister.BLOOD.get(), new ItemRenderingParams(new FixedRGB((short) 255, (short) 100, (short) 50),
                     GlintRenderTypes.createDirectGlint("direct_darkness", new ResourceLocation("pride:textures/glints/darkness_glint.png")),
                     GlintRenderTypes.createDirectEntityGlint("direct_entity_darkness", new ResourceLocation("pride:textures/glints/darkness_glint.png")))) {
