+++ src/main/java/com/robson/pride/api/elements/IceElement.java
@@ -18,7 +18,7 @@ import net.minecraft.world.entity.Entity;
 public interface IceElement {
 
     ElementData DATA = new ElementData("Ice", ServerDataManager.ICE, ParticleRegistry.SNOWFLAKE_PARTICLE.get(), ChatFormatting.DARK_AQUA,
-            SoundRegistry.CONE_OF_COLD_LOOP.get(), (byte) 5, SchoolRegister.ICE.get(), new ItemRenderingParams(new FixedRGB(50, 100, 250),
+            SoundRegistry.CONE_OF_COLD_LOOP.get(), (byte) 5, SchoolRegister.ICE.get(), new ItemRenderingParams(new FixedRGB((short) 50, (short) 100, (short) 250),
             GlintRenderTypes.createDirectGlint("direct_darkness", new ResourceLocation("pride:textures/glints/darkness_glint.png")),
             GlintRenderTypes.createDirectEntityGlint("direct_entity_darkness", new ResourceLocation("pride:textures/glints/darkness_glint.png")))) {
