+++ src/main/java/com/robson/pride/api/utils/math/FixedRGB.java
@@ -1,10 +1,10 @@
 package com.robson.pride.api.utils.math;
 
-public record FixedRGB(int r, int g, int b) {
+public record FixedRGB(short r, short g, short b) {
 
-    public FixedRGB(int r, int g, int b) {
-        this.r = 255 - r;
-        this.g = 255 - g;
-        this.b = 255 - b;
+    public FixedRGB(short r, short g, short b) {
+        this.r = (short) (255 - r);
+        this.g = (short) (255 - g);
+        this.b = (short) (255 - b);
     }
 }
